Add WeatherForecastQuery class for filtering and sorting weather forecasts
